+++ CHANGELOG.md
@@ -20,6 +20,10 @@ https://keepachangelog.com/ja/1.0.0/ に基づいて記述していく
 Unreleased
 -----------
 
+### Fixed
+- remove unnecessary omp_get_wtime() https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/432
+- fix logger bug in equations https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/432
+
 ### Changed
 - Start developing 0.15.3 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/431